% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakpicking.R
\name{helperpeak}
\alias{helperpeak}
\title{helper function for small peak elimination}
\usage{
helperpeak(thepos, vec, nsd, npos)
}
\arguments{
\item{thepos}{integer position of peak in vector vec}

\item{vec}{vector of values for peakdetection}

\item{nsd}{numeric minimum number of standard deviations for a peak to rise
above the mean of its immediate vicinity in order to be considered for a
peak call}

\item{npos}{integer value, peak size will be estimated plus/minus npos
positions from peak}
}
\value{
boolean TRUE if the peak at position thepos is to be deleted
}
\description{
helper function for small peak elimination
}
\keyword{internal}

