% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakpicking.R
\name{keepmax}
\alias{keepmax}
\title{Eliminates peak pairs too close to each other.}
\usage{
keepmax(vec, candidates, neighlim)
}
\arguments{
\item{vec}{vector of values}

\item{candidates}{boolean vector representing candidate peaks}

\item{neighlim}{integer limit of how far apart peaks must be}
}
\value{
boolean vector representing candidate peaks with smaller peaks of all
  pairs within neighlim removed
}
\description{
Eliminates peak pairs too close to each other.
}
\keyword{internal}

