% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakpicking.R
\name{smallpeaks}
\alias{smallpeaks}
\title{delete small peaks from vector with peak candidates}
\usage{
smallpeaks(vec, candidates, nsd, npos)
}
\arguments{
\item{vec}{vector of values for peak detection}

\item{candidates}{boolean vector of peak candidates}

\item{nsd}{numeric minimum number of standard deviations for a peak to rise
above the mean of its immediate vicinity in order to be considered for a
peak call}

\item{npos}{integer value, peak size will be estimated plus/minus npos
positions from peak}
}
\value{
boolean vector of peak candidates with small peaks eliminated
}
\description{
delete small peaks from vector with peak candidates
}
\keyword{internal}

